added $help command

+++ bot.py
@@ -2,14 +2,19 @@
 import discord, sys, re, os
 from commands import *
 
+sys.path.insert(0, '.')
+
 if len(sys.argv) != 2 :
     print("python bot.py <TOKEN.txt>")
     sys.exit(1)
 
 client = discord.Client()
 
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+
 textFiles = os.listdir("./pasta")
 commandFiles = os.listdir("./commands")
+
 if "__init__.py" in commandFiles:
     del commandFiles[commandFiles.index("__init__.py")]
 
@@ -24,32 +29,41 @@ for i in range(len(textFiles)):
     fileName = textFiles[i]
     textFiles[i] = fileName[0:len(fileName)-4:1]
 
+allowedChannels = [749388765717332019]
+
+stuff = {
+"message": "",
+"textFiles": textFiles,
+"commandFiles": commandFiles
+}
 
 @client.event
 async def on_message(message):
     print('\t' + message.author.display_name + ":\n" + message.content)
 
-    #commands
-    if message.content.startswith("$"):
-        command = message.content[1:len(message.content):1]
-        if command in commandFiles:
-            try:
-                await eval(command + '.' + command + "(message)")
-            except:
-                pass
-        else:
-            print("invalid command : " + command)
-    #text
-    elif message.content.startswith('!'):
-        command = message.content[1:len(message.content):1]
-        if command in textFiles:
-            try:
-                await message.delete()
-                f = open("pasta/" + command + ".txt", 'r', encoding="utf-8")
-                await message.channel.send(f.read())
-                f.close()
-            except:
-                pass
+    if message.channel.id in allowedChannels and not message.author.bot:
+        #commands
+        if message.content.startswith("$"):
+            command = message.content[1:len(message.content):1]
+            if command in commandFiles:
+                try:
+                    stuff["message"] = message
+                    await eval(command + '.' + command + "(stuff)")
+                except:
+                    pass
+            else:
+                print("invalid command : " + command)
+        #text
+        elif message.content.startswith('!'):
+            command = message.content[1:len(message.content):1]
+            if command in textFiles:
+                try:
+                    await message.delete()
+                    f = open("pasta/" + command + ".txt", 'r', encoding="utf-8")
+                    await message.channel.send(f.read())
+                    f.close()
+                except:
+                    pass
 
 @client.event
 async def on_reaction_add(reaction, user):
